Fix Fault Tolerance multiple metrics provider test Split OTel related configuration APIs to OpenTelemetryServerCOnfiguration
